refactor test_variational to move out the finite_diff utilities
